requirements-dev.txt
# Development dependencies for HealthEconomics360
# Install with: pip install -r requirements-dev.txt

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Security scanning
bandit>=1.7.5
safety>=2.3.0

# Testing and coverage
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-flask>=1.2.0

# Development tools
pre-commit>=3.3.0
python-dotenv>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0